Assorted dependency updates
These updates fix building the project on my up to date Linux box due to
missing cstdint headers on the older Draco versions it was pinned to,
and bumps Conan 1.x to the latest semver-compatible release, which also
introduces support for newer GCC compilers.

While at it, I've also bumped the GitHub Actions checkout version tag to
v4, from v3, and the runner image versions. The Ubuntu runner version
was purposefully not updated to avoid the risk of making binaries less
portable than before due to a bump in the `glibc` version that the
compiler will link to.